#35 create apps directory and move codes into it
